% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{assoc_plot_save}
\alias{assoc_plot_save}
\title{assoc_plot_save}
\usage{
assoc_plot_save(x, file, width = 9, height = 7, dpi=500)
}
\arguments{
\item{x}{the plot}

\item{file}{the filepath}

\item{width}{the width of the plot in inches}

\item{height}{the height of the plot in inches}

\item{dpi}{the resolution of the plot}
}
\description{
assoc_plot_save saves a png of the assoc_plot with the correct dimensions
}
\author{
James R Staley <jrstaley95@gmail.com>
}
