% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{stack_assoc_plot}
\alias{stack_assoc_plot}
\title{stack_assoc_plot}
\usage{
stack_assoc_plot(markers, z, corr=NULL, corr.top=NULL, traits, ylab=NULL,
  type="log10p", x.min = NULL, x.max = NULL, top.marker = NULL, legend=TRUE,
  build=37, highlights=NULL)
}
\arguments{
\item{markers}{data.frame of markers with markername (marker), chromosome (chr) and position (pos)}

\item{z}{matrix of Z-scores or probabilities with one column for each trait}

\item{corr}{(Pearson) correlation matrix between markers (i.e. r not r2)}

\item{corr.top}{(Pearson) correlation statistics between the top marker and the rest of the markers (i.e. r not r2)}
  
\item{traits}{trait names}

\item{ylab}{the y-axis label}

\item{type}{the type of the plot either log10p or probabilities}
  
\item{x.min}{start of region}

\item{x.max}{end of region}

\item{top.marker}{the marker to be annotated in the plot}
  
\item{legend}{add r2 legend}

\item{build}{genome build}

\item{highlights}{additional points to highlight}
}
\description{
stack_assoc_plot plots stacked regional association plots
}
\author{
James R Staley <jrstaley95@gmail.com>
}
