% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{stack_assoc_plot_save}
\alias{stack_assoc_plot_save}
\title{stack_assoc_plot_save}
\usage{
stack_assoc_plot_save(x, file, n_traits, width = NULL, height = NULL, dpi=500)
}
\arguments{
\item{x}{the plot}

\item{file}{the filepath}

\item{n_traits}{the filepath}

\item{width}{the width of the plot in inches}

\item{height}{the height of the plot in inches}

\item{dpi}{the resolution of the plot}
}
\description{
stack_assoc_plot_save saves a png of the assoc_plot with the correct dimensions
}
\author{
James R Staley <jrstaley95@gmail.com>
}
