% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{assoc_plot}
\alias{assoc_plot}
\title{assoc_plot}
\usage{
assoc_plot(data, corr=NULL, corr.top=NULL, ylab = NULL, title = NULL,
  subtitle = NULL, type = "log10p", x.min = NULL, x.max = NULL,
  top.marker=NULL, legend=TRUE)
}
\arguments{
\item{data}{data.frame with markername (marker), chromosome (chr), build 37 position (pos) and either z-statistics (z) or probabilities (prob) columns}

\item{corr}{(Pearson) correlation matrix between markers (i.e. r not r2)}

\item{corr.top}{(Pearson) correlation statistics between the top marker and the rest of the markers (i.e. r not r2)}
  
\item{ylab}{the y-axis label}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{type}{the type of the plot either log10p or probabilities}

\item{x.min}{start of region}

\item{x.max}{end of region}

\item{top.marker}{the top associated marker, i.e. the marker with the largest -log10p or probability}
  
\item{legend}{add r2 legend}
}
\description{
assoc_plot plots a scatter graph of associations (e.g. log10 p-values)
}
\author{
James R Staley <jrstaley95@gmail.com>
}
